refactor: move router in separate file
